Test the src, dev, and prod builds
I wanted to run some sanity check tests against the built versions as
well as the original source.

To do taht was pretty simple: I just required the `build/*` files from
the `GeneralStore-integration-test` and ran the "test" function against
each.

The only detail is that we need to run the build before running the unit
tests.

so `bin/build-and-test` now executes in the following order:

1. `flow check`
2. `bin/build-dev`
3. `bin/build-prod`
4. `jest`