Clear dispatch token (#52)
It seems that in certain versions of React it was possible for
`componentWillMount` to be called more than once. https://github.com/facebook/react/pull/6613

If that happened to a `connect`ed component, it would cause
`dispatcher.unregister` to be called again with a token that was already
unregistered causing an exception.

This `null`s the token field so unregister can only be called once for a
given token.